requirements.txt
streamlit>=1.20.0
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
matplotlib>=3.7.1
seaborn>=0.12.2